+++ src/main/java/bowling/domain/Frame.java
@@ -2,23 +2,8 @@ package bowling.domain;
 
 import bowling.FrameStateRenderer;
 
-public class Frame {
-
-    private FrameState state;
-
-    public Frame() {
-        this.state = new Ready();
-    }
-
-    public void bowl(PinCount fallenPinCount) {
-        state = state.bowl(fallenPinCount);
-    }
-
-    public boolean isFinished() {
-        return state instanceof Finished;
-    }
-
-    public FrameStateRenderer toFrameStateRenderer() {
-        return state.toRenderer();
-    }
+public interface Frame {
+    void bowl(PinCount fallenPinCount);
+    boolean isFinished();
+    FrameStateRenderer toFrameStateRenderer();
 }
