+++ src/main/java/bowling/domain/Frames.java
@@ -12,7 +12,7 @@ public class Frames {
     private final List<Frame> frames;
 
     public Frames() {
-        frames = Stream.generate(Frame::new)
+        frames = Stream.generate(NormalFrame::new)
                 .limit(FrameNumber.LAST_FRAME_NUMBER)
                 .collect(Collectors.collectingAndThen(Collectors.toList(), Collections::unmodifiableList));
     }
